% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcUtils.R
\docType{package}
\name{bmcUtils}
\alias{bmcUtils}
\alias{bmcUtils-package}
\alias{NumberOfDays}
\alias{StripDegMinSec}
\alias{UnlistDate}
\alias{PullFilenameDates}
\alias{LoopStatus}
\alias{SnipSingleCharacter}
\alias{Multiplot}
\title{Spare utilities package}
\usage{
NumberOfDays(date)

StripDegMinSec(x)

UnlistDate(x)

PullFilenameDates(fmt)

LoopStatus(from, to, big_inc = 1000, digits = 1)

SnipSingleCharacter(v, side = "front")

Multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
This is a odds-and-ends repository for generalized functions I use in developing
other R packages or analyses. Sometimes other packages are required for functions
to work, but I dont include them in the DESCRIPTION to make installation easier.
}
\section{Functions}{
\itemize{
\item \code{NumberOfDays}: Counts days in a month

\item \code{StripDegMinSec}: Strips deg-min-sec vectors to 'numeric' char vector.

\item \code{UnlistDate}: Unlists a list of 'Date' objects w/o coercion to numeric.

\item \code{PullFilenameDates}: Generates a date vector from filenames.

\item \code{LoopStatus}: Prints loop progress.

\item \code{SnipSingleCharacter}: Snips one character from each element of a character vector

\item \code{Multiplot}: Multiple ggplot panel
}}

\section{Functions}{


`NumberOfDays` counts the days in a month. Input is an object of class
`date`, and it will work for leap years etc.

`StripDegMinSec` will convert a lat/long degree-min-sec character vector
to a 'numeric' character vector, where each character is three numeric
values seperated by single spaces. This makes the vector usable by
the `conv_units` function in the `measurements` package >= v. 1.2.0.

`UnlistDate` will unlist a list of 'Date' objects without coercing them to
numeric.

`PullFilenameDates` extracts dates of the given format from the current
working directory. Assumes that the format doesn't include alphabetic
characters.

`LoopStatus` prints the progress of a slow looop.

`SnipSingleCharacter` trims a character vector by 1 character for each vector element.

`Multiplot` plots multiple ggplots in the same window - its mostly copied
verbatim from the R cookbook, link in the function source.
}

\author{
Brandon McNellis
}
