% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{NumberOfDays}
\alias{StripDegMinSec}
\alias{UnlistDate}
\alias{PullFilenameDates}
\alias{clear}
\alias{LoopStatus}
\alias{SnipSingleCharacter}
\title{Internal utility functions.}
\usage{
utils()

NumberOfDays(date)

StripDegMinSec(x)

UnlistDate(x)

PullFilenameDates(fmt)

clear()

LoopStatus(from, to, digits = 1)

SnipSingleCharacter(v, side = "front")
}
\description{
`NumberOfDays` counts the days in a month. Input is an object of class
`date`, and it will work for leap years etc.

`StripDegMinSec` will convert a lat/long degree-min-sec character vector
to a 'numeric' character vector, where each character is three numeric
values seperated by single spaces. This makes the vector usable by
the `conv_units` function in the `measurements` package >= v. 1.2.0.

`UnlistDate` will unlist a list of 'Date' objects without coercing them to
numeric.

`PullFilenameDates` extracts dates of the given format from the current
working directory. Assumes that the format doesn't include alphabetic
characters.

`clear` is a shortcut function that (1) clears the environment, (2) runs GC,
(3) clears the console, and (4) restarts R.

`LoopStatus` prints the progress of a slow looop.

`SnipSingleCharacter` trims a character vector by 1 character for each vector element.
}
\section{Functions}{
\itemize{
\item \code{NumberOfDays}: Counts days in a month

\item \code{StripDegMinSec}: Strips deg-min-sec vectors to 'numeric' char vector.

\item \code{UnlistDate}: Unlists a list of 'Date' objects w/o coercion to numeric.

\item \code{PullFilenameDates}: Generates a date vector from filenames.

\item \code{clear}: Clears multiple aspects of the R environment.

\item \code{LoopStatus}: Prints loop progress.

\item \code{SnipSingleCharacter}: Snips one character from each element of a character vector
}}

\examples{
NumberOfDays(Sys.Date())
}
